numeric extension move into core 1
moved numeric/algorithm.hpp into core
