+++ example/adaptive_histogram_equalization.cpp
@@ -1,5 +1,6 @@
 //
 // Copyright 2020 Debabrata Mandal <mandaldebabrata123@gmail.com>
+// Copyright 2021 Pranam Lashkari <plashkari628@gmail.com>
 //
 // Use, modification and distribution are subject to the Boost Software License,
 // Version 1.0. (See accompanying file LICENSE_1_0.txt or copy at
@@ -8,7 +9,7 @@
 
 #include <boost/gil.hpp>
 #include <boost/gil/extension/io/png.hpp>
-#include <boost/gil/extension/numeric/algorithm.hpp>
+#include <boost/gil/algorithm.hpp>
 #include <boost/gil/image_processing/adaptive_histogram_equalization.hpp>
 
 using namespace boost::gil;
