+++ include/boost/gil/extension/numeric/convolve.hpp
@@ -1,6 +1,7 @@
 //
 // Copyright 2005-2007 Adobe Systems Incorporated
 // Copyright 2019 Miral Shah <miralshah2211@gmail.com>
+// Copyright 2021 Pranam Lashkari <plashkari628@gmail.com>
 //
 // Distributed under the Boost Software License, Version 1.0
 // See accompanying file LICENSE_1_0.txt or copy at
@@ -9,7 +10,6 @@
 #ifndef BOOST_GIL_EXTENSION_NUMERIC_CONVOLVE_HPP
 #define BOOST_GIL_EXTENSION_NUMERIC_CONVOLVE_HPP
 
-#include <boost/gil/extension/numeric/algorithm.hpp>
 #include <boost/gil/extension/numeric/kernel.hpp>
 #include <boost/gil/extension/numeric/pixel_numeric_operations.hpp>
