+++ test/core/algorithm/CMakeLists.txt
@@ -1,5 +1,6 @@
 #
 # Copyright (c) 2018 Mateusz Loskot <mateusz at loskot dot net>
+# Copyright (c) 2021 Pranam Lashkari <plashkari628@gmail.com>
 #
 # Distributed under the Boost Software License, Version 1.0.
 # (See accompanying file LICENSE_1_0.txt or copy at
@@ -8,7 +9,8 @@
 foreach(_name
   for_each_pixel
   std_fill
-  std_uninitialized_fill)
+  std_uninitialized_fill
+  extend_boundary)
   set(_test t_core_algorithm_${_name})
   set(_target test_core_algorithm_${_name})
