+++ test/core/algorithm/Jamfile
@@ -1,6 +1,7 @@
 # Boost.GIL (Generic Image Library) - tests
 #
 # Copyright (c) 2018-2019 Mateusz Loskot <mateusz at loskot dot net>
+# Copyright (c) 2021 Pranam Lashkari <plashkari628@gmail.com>
 #
 # Distributed under the Boost Software License, Version 1.0.
 # (See accompanying file LICENSE_1_0.txt or
@@ -11,3 +12,4 @@ import testing ;
 run for_each_pixel.cpp ;
 run std_fill.cpp ;
 run std_uninitialized_fill.cpp ;
+run extend_boundary.cpp ;
