+++ test/core/algorithm/extend_boundary.cpp
@@ -1,13 +1,14 @@
 //
-// Copyright 2019 Pranam Lashkari <plashkari628@gmail.com>
+// Copyright 2019-2021 Pranam Lashkari <plashkari628@gmail.com>
 //
 // Distributed under the Boost Software License, Version 1.0
 // See accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt
 //
 #include <boost/gil.hpp>
-#include <boost/gil/extension/numeric/algorithm.hpp>
 #include <boost/gil/extension/numeric/convolve.hpp>
+
+#include <boost/gil/algorithm.hpp>
 
 #include <boost/core/lightweight_test.hpp>
