+++ test/extension/numeric/CMakeLists.txt
@@ -1,5 +1,6 @@
 #
 # Copyright (c) 2017 Mateusz Loskot <mateusz at loskot dot net>
+# Copyright (c) 2021 Pranam Lashkari <plashkari628@gmail.com>
 # All rights reserved.
 #
 # Distributed under the Boost Software License, Version 1.0.
@@ -14,7 +15,6 @@ foreach(_name
   convolve_2d
   convolve_cols
   convolve_rows
-  extend_boundary
   kernel
   kernel_fixed
   matrix3x2
