+++ test/extension/numeric/Jamfile
@@ -2,6 +2,7 @@
 #
 # Copyright (c) 2013 Christian Henning
 # Copyright (c) 2019 Mateusz Loskot <mateusz at loskot dot net>
+# Copyright (c) 2021 Pranam Lashkari <plashkari628@gmail.com>
 #
 # Distributed under the Boost Software License, Version 1.0.
 # (See accompanying file LICENSE_1_0.txt or
@@ -18,7 +19,6 @@ run convolve.cpp ;
 run convolve_2d.cpp ;
 run convolve_cols.cpp ;
 run convolve_rows.cpp ;
-run extend_boundary.cpp ;
 run kernel.cpp ;
 run kernel_fixed.cpp ;
 run matrix3x2.cpp ;
